+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2
+
+- Added `parseSrecView` and `toSrecView` methods to handle SREC view
+- Added `view2srec` command line utility to convert from SREC view to SREC file
+
 ## 0.0.1
 
 - Initial version, created by Stagehand